chore: update test command
