+++ frontend/desktop/scripts/getMongodbUri.ps1
@@ -1,0 +1,5 @@
+$_mongodbUri = (kubectl.exe get secret/desktop-frontend-secret -nsealos -o json | 
+ConvertFrom-Json).data.mongodb_uri
+Write-Host $_mongodbUri
+$mongodbUri = ConvertFrom-Base64ToString $_mongodbUri
+Write-Host $mongodbUri
