+++ frontend/desktop/src/__tests__/api/e2e/namespace/withoutSession.test.ts
